Move concrete KnowledgeGroup into its own file
